add_flow_type
