+++ src/promptflow-azure/promptflow/azure/operations/_run_operations.py
@@ -1124,3 +1124,15 @@ class RunOperations(WorkspaceTelemetryMixin, _ScopeDependentOperations):
             run._identity[IdentityKeys.RESOURCE_ID] = resource_id
         else:
             raise UserErrorException(f"Identity type {identity_type!r} is not supported.")
+
+        def _get_telemetry_values(self, *args, **kwargs):
+            activity_name = kwargs.get("activity_name", None)
+            telemetry_values = super()._get_telemetry_values(*args, **kwargs)
+            try:
+                if activity_name == "pfazure.runs.create_or_update":
+                    run: Run = kwargs.get("run", None) or args[0]
+                    telemetry_values["flow_type"] = run._flow_type
+            except Exception as e:
+                logger.error(f"Failed to get telemetry values: {str(e)}")
+
+            return telemetry_values
